Add the tutorials in the Resources, general cleanup.
Everywhere:
* Created a footer, with button-like links to GitHub and Travis.
* Some links to precice.org pages contained the domain name. I converted these to relative, so that they are not prune to an address change and that they work in a local, offline Jekyll installation.
* Some trailing spaces were automatically removed.

Feature page:
* MAC --> macOS
* Fortran90 --> Fortran 90
* Endpoints in full sentences

Resources:
* Removed the build status (now moved to footer).
* Moved the branch information next to the repository link.
* Changed the link to v1.0.0 to a link to the always latest release.
* Created a table with the existing tutorials, with images. 

Publications:
* Placed the "to cite preCICE, please use" message also here.
* Added Chourdakis' thesis and conference paper for adapters for open-source solvers.

About:
* Corrected my name to the official, full variant and updated my status.

The GitHub button requires a small a small script, which is placed at the end of the page, so that it does not delay the content (default.HTML). The respective HTML is inserted in the footer.html. The code was generated from buttons.github.io and edited for the name.

The footer is set to at least 80px height. Having a footer also helps while reading.

The images in the tutorial also follow  similar responsive rules with the ones set for the testimonials.