Replace the hard-coded news with an automatic approach.
This adds a Javascript script in the footer, which converts the GitHub Atom feeds for the preCICE releases: https://github.com/precice/precice/releases.atom

to a json file using the Yahoo API, extracts the latest entry and formats it in exactly the same format as before.

No need to update the news manually anymore!

While this is written to show only the latest entry, it can be easily modified to show more entries from the same feed.